+++ inst/lua/inline-code.lua
@@ -1,0 +1,11 @@
+function Code(elem)
+  if elem.attributes['font.family'] == nil then
+    elem.attributes['font.family'] = 'monospace'
+  end
+  if elem.attributes['font.family'] == nil then
+    elem.attributes['shading.color'] = '#f8f8f8'
+  end
+  local span = pandoc.Span(elem.text)
+  span.attr = elem.attr
+  return span
+end